bsr(api): move delete learner to learner-management bounded context
