+++ api/lib/routes.js
@@ -36,7 +36,7 @@ import * as organizationInvitations from './application/organization-invitations
 import * as organizations from './application/organizations/index.js';
 import * as organizationsAdministration from './application/organizations-administration/index.js';
 import * as organizationLearners from './application/organization-learners/index.js';
-import * as organizationLearnersManagement from './application/organization-learners-management/index.js';
+
 import * as passwords from './application/passwords/index.js';
 import * as poleEmploi from './application/pole-emploi/index.js';
 import * as prescribers from './application/prescribers/index.js';
@@ -96,7 +96,6 @@ const routes = [
   lcms,
   memberships,
   organizationLearners,
-  organizationLearnersManagement,
   organizations,
   organizationsAdministration,
   passwords,
