+++ api/src/prescription/learner-management/application/organization-learners-route.js
@@ -1,6 +1,6 @@
-import { securityPreHandlers } from '../security-pre-handlers.js';
+import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
 import Joi from 'joi';
-import { identifiersType } from '../../domain/types/identifiers-type.js';
+import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
 import { organizationLearnersController } from './organization-learners-controller.js';
 
 const register = async (server) => {
