+++ api/src/prescription/learner-management/routes.js
@@ -1,5 +1,6 @@
 import * as supOrganizationManangement from './application/sup-organization-management-route.js';
+import * as organizationLearners from './application/organization-learners-route.js';
 
-const supOrganizationManangementRoutes = [supOrganizationManangement];
+const learnerManagementRoutes = [supOrganizationManangement, organizationLearners];
 
-export { supOrganizationManangementRoutes };
+export { learnerManagementRoutes };
