+++ api/tests/prescription/learner-management/acceptance/application/organization-controller-replace-sup-organization-learners_test.js
@@ -1,8 +1,8 @@
-import { expect, knex, databaseBuilder, generateValidRequestAuthorizationHeader } from '../../../test-helper.js';
-import { Membership } from '../../../../lib/domain/models/Membership.js';
-import { SupOrganizationLearnerImportHeader } from '../../../../src/prescription/learner-management/infrastructure/serializers/csv/sup-organization-learner-import-header.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
-import { createServer } from '../../../../server.js';
+import { expect, knex, databaseBuilder, generateValidRequestAuthorizationHeader } from '../../../../test-helper.js';
+import { Membership } from '../../../../../lib/domain/models/Membership.js';
+import { SupOrganizationLearnerImportHeader } from '../../../../../src/prescription/learner-management/infrastructure/serializers/csv/sup-organization-learner-import-header.js';
+import { getI18n } from '../../../../tooling/i18n/i18n.js';
+import { createServer } from '../../../../../server.js';
 
 const i18n = getI18n();
 const supOrganizationLearnerImportHeader = new SupOrganizationLearnerImportHeader(i18n).columns
