+++ api/tests/prescription/learner-management/integration/infrastructure/repositories/campaign-participation-repository_test.js
@@ -1,6 +1,6 @@
-import { expect, knex, databaseBuilder, sinon } from '../../../../test-helper.js';
-import { removeByOrganizationLearnerIds } from '../../../../../lib/infrastructure/repositories/organization-learners-management/campaign-participation-repository.js';
-import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
+import { expect, knex, databaseBuilder, sinon } from '../../../../../test-helper.js';
+import { removeByOrganizationLearnerIds } from '../../../../../../src/prescription/learner-management/infrastructure/repositories/campaign-participation-repository.js';
+import { DomainTransaction } from '../../../../../../lib/infrastructure/DomainTransaction.js';
 
 describe('Integration | Repository | Organization Learners Management | Campaign Participation', function () {
   describe('#removeByOrganizationLearnerIds', function () {
